Add files via upload
A similar algorithm to Bellman-Ford to find shortest paths from a source vertex to every other vertex. But it uses a queue data structure and somewhat intuitive than Bellman-Ford. Rather than going through every edge in each iteration, it uses queue to keep the only nodes whose shortest distance is relabelled and then consider the adjacent edges of those present in the queue.