Create genericSwap.c
This code swaps two variables. But the function in code can be used to swap any data type like int,float,double,char...but when dealing with this type of generic function we should be careful as it can be messed up easily. Generic functions are somewhat better than Template as a single function code is only compiled and same assembly code is used is for different data types rather than separate copies for each data types in case of Template. But one should be careful as using generic codes we convince compiler that everything is fine so it does not give potential errors or warnings like in normal cases.